+++ importlib_metadata/tests/test_zip.py
@@ -2,8 +2,6 @@ import sys
 import unittest
 import importlib_metadata
 
-from types import ModuleType
-
 try:
     from contextlib import ExitStack
 except ImportError:
@@ -36,25 +34,9 @@ class TestZip(unittest.TestCase):
         entry_point = parser.get('console_scripts', 'example')
         self.assertEqual(entry_point, 'example:main')
 
-    def test_not_a_zip(self):
-        # For coverage purposes, this module is importable, but has neither a
-        # location on the file system, nor a .archive attribute.
-        sys.modules['bespoke'] = ModuleType('bespoke')
-        self.resources.callback(sys.modules.pop, 'bespoke')
-        self.assertRaises(ImportError,
-                          importlib_metadata.version,
-                          'bespoke')
-
-    def test_unversioned_dist_info(self):
-        # For coverage purposes, give the module an unversioned .archive
-        # attribute.
-        bespoke = sys.modules['bespoke'] = ModuleType('bespoke')
-        bespoke.__loader__ = BespokeLoader()
-        self.resources.callback(sys.modules.pop, 'bespoke')
-        self.assertRaises(ImportError,
-                          importlib_metadata.version,
-                          'bespoke')
-
     def test_missing_metadata(self):
         distribution = importlib_metadata.distribution('example')
         self.assertIsNone(distribution.read_text('does not exist'))
+
+    def test_case_insensitive(self):
+        self.assertEqual(importlib_metadata.version('Example'), '21.12')